% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualization_networks.R
\name{add_component_information}
\alias{add_component_information}
\title{Add information component data to nodes}
\usage{
add_component_information(network_data, universal_info)
}
\description{
Add information component data to nodes
}
\keyword{internal}
