% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/consensus_diversity.R
\name{analyze_metric_conflicts}
\alias{analyze_metric_conflicts}
\title{Analyze conflicts between metrics}
\usage{
analyze_metric_conflicts(metric_data, groups, alpha)
}
\description{
Analyze conflicts between metrics
}
\keyword{internal}
