% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shapley_values.R
\name{assess_approximation_quality}
\alias{assess_approximation_quality}
\title{Assess Approximation Quality}
\usage{
assess_approximation_quality(shapley_results, method, n_samples)
}
\arguments{
\item{shapley_results}{Results from approximate method}

\item{method}{Approximation method used}

\item{n_samples}{Number of samples used}
}
\value{
Quality assessment metrics
}
\description{
Evaluates the quality of approximate Shapley value calculations
}
\keyword{internal}
