% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/universal_information.R
\name{assess_deconvolution_quality}
\alias{assess_deconvolution_quality}
\title{Assess quality of deconvolution}
\usage{
assess_deconvolution_quality(
  diversity_metrics,
  information_components,
  transformation_matrix
)
}
\description{
Assess quality of deconvolution
}
\keyword{internal}
