% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bootstrap_uncertainty.R
\name{assess_method_agreement}
\alias{assess_method_agreement}
\title{Assess Method Agreement}
\usage{
assess_method_agreement(bootstrap_results)
}
\arguments{
\item{bootstrap_results}{Bootstrap results for all methods}
}
\value{
Agreement assessment
}
\description{
Assesses agreement between indicator methods
}
\keyword{internal}
