% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/universal_transformations.R
\name{assess_prediction_quality}
\alias{assess_prediction_quality}
\title{Assess prediction quality}
\usage{
assess_prediction_quality(
  predicted_metrics,
  target_metrics,
  transformation_matrix
)
}
\description{
Assess prediction quality
}
\keyword{internal}
