% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format_converters.R
\name{biom_to_phyloseq}
\alias{biom_to_phyloseq}
\title{Convert BIOM file to phyloseq}
\usage{
biom_to_phyloseq(
  biom_file,
  tree_file = NULL,
  sample_metadata = NULL,
  parse_taxonomy = TRUE
)
}
\arguments{
\item{biom_file}{Path to BIOM file}

\item{tree_file}{Optional path to tree file (Newick format)}

\item{sample_metadata}{Optional data frame or path to sample metadata file}

\item{parse_taxonomy}{Logical, whether to parse taxonomy from observation metadata}
}
\value{
phyloseq object
}
\description{
Convert BIOM file to phyloseq
}
\examples{
\dontrun{
# Convert BIOM file
ps <- biom_to_phyloseq("otu_table.biom", 
                      tree_file = "tree.nwk",
                      sample_metadata = "sample_data.tsv")
}
}
