% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualization_networks.R
\name{build_network_from_transformation_matrix}
\alias{build_network_from_transformation_matrix}
\title{Build network data from transformation matrix}
\usage{
build_network_from_transformation_matrix(universal_info, min_r_squared)
}
\description{
Build network data from transformation matrix
}
\keyword{internal}
