% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cache_system.R
\name{cache_exists}
\alias{cache_exists}
\title{Check if object exists in cache}
\usage{
cache_exists(key)
}
\arguments{
\item{key}{Cache key}
}
\value{
Logical indicating if key exists in cache
}
\description{
Check if object exists in cache
}
