% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cache_system.R
\name{cache_get}
\alias{cache_get}
\title{Retrieve object from cache}
\usage{
cache_get(key, default = NULL)
}
\arguments{
\item{key}{Cache key}

\item{default}{Default value if not found in cache}
}
\value{
Cached object or default value
}
\description{
Retrieve object from cache
}
\examples{
\dontrun{
# Retrieve from cache
data <- cache_get("my_dataset")

# With default value
data <- cache_get("my_dataset", default = NULL)
}
}
