% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cache_system.R
\name{cache_is_initialized}
\alias{cache_is_initialized}
\title{Check if cache system is initialized}
\usage{
cache_is_initialized()
}
\value{
Logical indicating if cache is initialized
}
\description{
Check if cache system is initialized
}
