% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cache_system.R
\name{cache_remove}
\alias{cache_remove}
\title{Remove object from cache}
\usage{
cache_remove(key)
}
\arguments{
\item{key}{Cache key}
}
\value{
Logical indicating success
}
\description{
Remove object from cache
}
