% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shapley_values.R
\name{calculate_coalition_value}
\alias{calculate_coalition_value}
\title{Calculate Coalition Value}
\usage{
calculate_coalition_value(coalition_otu, comp_data, component)
}
\arguments{
\item{coalition_otu}{OTU matrix for coalition taxa}

\item{comp_data}{Component data matrix}

\item{component}{Component to calculate ("richness", "evenness", etc.)}
}
\value{
Coalition value
}
\description{
Calculates the value contributed by a specific coalition of taxa
}
\keyword{internal}
