% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/universal_information.R
\name{calculate_comprehensive_metrics}
\alias{calculate_comprehensive_metrics}
\title{Calculate comprehensive diversity metrics for deconvolution}
\usage{
calculate_comprehensive_metrics(
  phyloseq_object,
  groups,
  include_phylogenetic,
  include_spatial,
  ...
)
}
\description{
Calculate comprehensive diversity metrics for deconvolution
}
\keyword{internal}
