% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/information_theory.R
\name{calculate_conditional_entropy}
\alias{calculate_conditional_entropy}
\title{Calculate Conditional Entropy}
\usage{
calculate_conditional_entropy(y, x)
}
\arguments{
\item{y}{Target variable}

\item{x}{Conditioning variable}
}
\value{
Conditional entropy
}
\description{
Calculates H(Y|X) = sum_x P(x) * H(Y|X=x)
}
\keyword{internal}
