% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/information_theory.R
\name{calculate_conditional_joint_entropy}
\alias{calculate_conditional_joint_entropy}
\title{Calculate Conditional Joint Entropy}
\usage{
calculate_conditional_joint_entropy(x, y, z)
}
\arguments{
\item{x}{First variable}

\item{y}{Second variable}

\item{z}{Conditioning variable}
}
\value{
Conditional joint entropy
}
\description{
Calculates H(X,Y|Z) = sum_z P(z) * H(X,Y|Z=z)
}
\keyword{internal}
