% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/information_theory.R
\name{calculate_conditional_mutual_information_trio}
\alias{calculate_conditional_mutual_information_trio}
\title{Calculate Conditional Mutual Information for Three Variables}
\usage{
calculate_conditional_mutual_information_trio(x, y, z, method)
}
\arguments{
\item{x}{First variable}

\item{y}{Second variable}

\item{z}{Conditioning variable}

\item{method}{Estimation method}
}
\value{
Conditional mutual information
}
\description{
Calculates I(X;Y|Z) using the definition I(X;Y|Z) = H(X|Z) + H(Y|Z) - H(X,Y|Z)
}
\keyword{internal}
