% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/consensus_diversity.R
\name{calculate_correlation_consensus}
\alias{calculate_correlation_consensus}
\title{Calculate correlation-weighted consensus}
\usage{
calculate_correlation_consensus(metric_data, diversity_results, groups, alpha)
}
\description{
Calculate correlation-weighted consensus
}
\keyword{internal}
