% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/performance_optimization.R
\name{calculate_diversity_sparse}
\alias{calculate_diversity_sparse}
\title{Calculate diversity metrics from sparse matrix}
\usage{
calculate_diversity_sparse(sparse_otu, metrics, physeq)
}
\description{
Calculate diversity metrics from sparse matrix
}
\keyword{internal}
