% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/information_theory.R
\name{calculate_entropy}
\alias{calculate_entropy}
\title{Calculate Shannon Entropy}
\usage{
calculate_entropy(x)
}
\arguments{
\item{x}{Discrete variable}
}
\value{
Shannon entropy
}
\description{
Calculates Shannon entropy for discrete probability distribution
}
\keyword{internal}
