% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shapley_values.R
\name{calculate_exact_shapley_values}
\alias{calculate_exact_shapley_values}
\title{Calculate Exact Shapley Values}
\usage{
calculate_exact_shapley_values(taxa_names, char_function, verbose, parallel)
}
\arguments{
\item{taxa_names}{Names of taxa}

\item{char_function}{Characteristic function}

\item{verbose}{Print progress}

\item{parallel}{Use parallel processing}
}
\value{
List with Shapley matrix and coalition values
}
\description{
Computes exact Shapley values by enumerating all possible coalitions
}
\keyword{internal}
