% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/information_theory.R
\name{calculate_information_gain}
\alias{calculate_information_gain}
\title{Calculate Information Gain}
\usage{
calculate_information_gain(discrete_taxa, comp_data, method)
}
\arguments{
\item{discrete_taxa}{Discretized taxa data}

\item{comp_data}{Component data}

\item{method}{Estimation method}
}
\value{
Information gain matrix
}
\description{
Calculates information gain for each taxa-component relationship
}
\keyword{internal}
