% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/consensus_diversity.R
\name{calculate_majority_consensus}
\alias{calculate_majority_consensus}
\title{Calculate majority vote consensus}
\usage{
calculate_majority_consensus(metric_data, diversity_results, groups, alpha)
}
\description{
Calculate majority vote consensus
}
\keyword{internal}
