% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shapley_values.R
\name{calculate_marginal_shapley_values}
\alias{calculate_marginal_shapley_values}
\title{Calculate Marginal-Based Shapley Values}
\usage{
calculate_marginal_shapley_values(taxa_names, char_function, verbose)
}
\arguments{
\item{taxa_names}{Names of taxa}

\item{char_function}{Characteristic function}

\item{verbose}{Print progress}
}
\value{
List with Shapley matrix and coalition values
}
\description{
Fast approximation using only pairwise marginal contributions
}
\keyword{internal}
