% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/information_theory.R
\name{calculate_mutual_information}
\alias{calculate_mutual_information}
\title{Calculate Mutual Information Between Two Variables}
\usage{
calculate_mutual_information(x, y, method)
}
\arguments{
\item{x}{First variable (discrete)}

\item{y}{Second variable (discrete)}

\item{method}{Estimation method}
}
\value{
Mutual information value
}
\description{
Core mutual information calculation using various estimators
}
\keyword{internal}
