% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualization_networks.R
\name{calculate_node_sizes}
\alias{calculate_node_sizes}
\title{Calculate node sizes based on importance}
\usage{
calculate_node_sizes(network_data, node_size_by)
}
\description{
Calculate node sizes based on importance
}
\keyword{internal}
