% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/robust_taxa_indicators.R
\name{calculate_overall_reliability}
\alias{calculate_overall_reliability}
\title{Calculate Overall Reliability Score}
\usage{
calculate_overall_reliability(reliability)
}
\arguments{
\item{reliability}{Component reliability scores}
}
\value{
Numeric overall score
}
\description{
Calculate Overall Reliability Score
}
\keyword{internal}
