% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/taxa_indicators.R
\name{calculate_phylogenetic_uniqueness}
\alias{calculate_phylogenetic_uniqueness}
\title{Calculate Phylogenetic Uniqueness}
\usage{
calculate_phylogenetic_uniqueness(tree, taxa)
}
\arguments{
\item{tree}{Phylogenetic tree}

\item{taxa}{Taxa names}
}
\value{
Numeric vector of uniqueness scores
}
\description{
Calculates how unique each taxon is based on phylogenetic position
}
\keyword{internal}
