% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shapley_values.R
\name{calculate_sampling_shapley_values}
\alias{calculate_sampling_shapley_values}
\title{Calculate Sampling-Based Shapley Values}
\usage{
calculate_sampling_shapley_values(
  taxa_names,
  char_function,
  n_samples,
  verbose
)
}
\arguments{
\item{taxa_names}{Names of taxa}

\item{char_function}{Characteristic function}

\item{n_samples}{Number of samples to use}

\item{verbose}{Print progress}
}
\value{
List with Shapley matrix and coalition values
}
\description{
Approximates Shapley values using random coalition sampling
}
\keyword{internal}
