% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/taxa_drivers.R
\name{calculate_taxonomic_uniqueness}
\alias{calculate_taxonomic_uniqueness}
\title{Calculate Taxonomic Uniqueness}
\usage{
calculate_taxonomic_uniqueness(physeq)
}
\arguments{
\item{physeq}{A phyloseq object}
}
\value{
Numeric vector of uniqueness scores
}
\description{
Calculates taxonomic uniqueness when no phylogenetic tree is available
}
\keyword{internal}
