% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/consensus_diversity.R
\name{calculate_weighted_consensus}
\alias{calculate_weighted_consensus}
\title{Calculate weighted mean consensus}
\usage{
calculate_weighted_consensus(metric_data, diversity_results, groups, alpha)
}
\description{
Calculate weighted mean consensus
}
\keyword{internal}
