% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/relationship_analysis.R
\name{cluster_metrics_by_information}
\alias{cluster_metrics_by_information}
\title{Cluster metrics by information profiles}
\usage{
cluster_metrics_by_information(component_contributions)
}
\description{
Cluster metrics by information profiles
}
\keyword{internal}
