% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/information_theory.R
\name{compute_conditional_mi_matrix}
\alias{compute_conditional_mi_matrix}
\title{Compute Conditional MI Matrix}
\usage{
compute_conditional_mi_matrix(
  discrete_taxa,
  discrete_comp,
  conditioning_taxa,
  target_taxa,
  method,
  verbose
)
}
\arguments{
\item{discrete_taxa}{Discretized taxa matrix}

\item{discrete_comp}{Discretized component matrix}

\item{conditioning_taxa}{Conditioning taxa names}

\item{target_taxa}{Target taxa names}

\item{method}{Estimation method}

\item{verbose}{Print progress}
}
\value{
Conditional MI matrix
}
\description{
Compute Conditional MI Matrix
}
\keyword{internal}
