% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/information_theory.R
\name{compute_mutual_information_matrix}
\alias{compute_mutual_information_matrix}
\title{Compute Mutual Information Matrix}
\usage{
compute_mutual_information_matrix(discrete_taxa, comp_data, method, verbose)
}
\arguments{
\item{discrete_taxa}{Discretized taxa abundance matrix}

\item{comp_data}{Component data matrix}

\item{method}{Information estimation method}

\item{verbose}{Print progress}
}
\value{
Matrix of mutual information values
}
\description{
Calculates mutual information between each taxon and each component
}
\keyword{internal}
