% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/universal_information.R
\name{compute_transformation_matrix}
\alias{compute_transformation_matrix}
\title{Compute transformation matrix between all diversity metrics}
\usage{
compute_transformation_matrix(diversity_metrics, information_components)
}
\description{
Compute transformation matrix between all diversity metrics
}
\keyword{internal}
