% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/information_theory.R
\name{compute_unconditional_mi_for_comparison}
\alias{compute_unconditional_mi_for_comparison}
\title{Compute Unconditional MI for Comparison}
\usage{
compute_unconditional_mi_for_comparison(
  discrete_taxa,
  discrete_comp,
  target_taxa,
  method
)
}
\arguments{
\item{discrete_taxa}{Discretized taxa matrix}

\item{discrete_comp}{Discretized component matrix}

\item{target_taxa}{Target taxa names}

\item{method}{Estimation method}
}
\value{
Unconditional MI matrix
}
\description{
Compute Unconditional MI for Comparison
}
\keyword{internal}
