% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bootstrap_uncertainty.R
\name{create_bootstrap_phyloseq}
\alias{create_bootstrap_phyloseq}
\title{Create Bootstrap Phyloseq Object}
\usage{
create_bootstrap_phyloseq(boot_otu, original_physeq)
}
\arguments{
\item{boot_otu}{Bootstrap OTU matrix}

\item{original_physeq}{Original phyloseq object}
}
\value{
Bootstrap phyloseq object
}
\description{
Creates phyloseq object from bootstrap OTU matrix
}
\keyword{internal}
