% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shapley_values.R
\name{create_characteristic_function}
\alias{create_characteristic_function}
\title{Create Characteristic Function for Coalition Game}
\usage{
create_characteristic_function(otu_subset, comp_data, verbose)
}
\arguments{
\item{otu_subset}{OTU matrix subset (selected taxa x samples)}

\item{comp_data}{Component data matrix (samples x components)}

\item{verbose}{Print progress}
}
\value{
Function that calculates coalition values
}
\description{
Defines the value of each possible coalition of taxa for Shapley calculation
}
\keyword{internal}
