% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualization_dashboards.R
\name{create_component_correlation_plot}
\alias{create_component_correlation_plot}
\title{Create component correlation plot}
\usage{
create_component_correlation_plot(components)
}
\description{
Create component correlation plot
}
\keyword{internal}
