% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualization_dashboards.R
\name{create_component_heatmap}
\alias{create_component_heatmap}
\title{Create component values heatmap}
\usage{
create_component_heatmap(components, color_scheme)
}
\description{
Create component values heatmap
}
\keyword{internal}
