% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualization_dashboards.R
\name{create_component_importance_plot}
\alias{create_component_importance_plot}
\title{Create component importance plot}
\usage{
create_component_importance_plot(component_importance)
}
\description{
Create component importance plot
}
\keyword{internal}
