% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualization_dashboards.R
\name{create_component_proportions_plot}
\alias{create_component_proportions_plot}
\title{Create component proportions stacked bar chart}
\usage{
create_component_proportions_plot(components, color_scheme)
}
\description{
Create component proportions stacked bar chart
}
\keyword{internal}
