% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/information_theory.R
\name{create_conditioning_context}
\alias{create_conditioning_context}
\title{Create Conditioning Context}
\usage{
create_conditioning_context(conditioning_matrix)
}
\arguments{
\item{conditioning_matrix}{Matrix of conditioning taxa (taxa x samples)}
}
\value{
Vector representing combined context
}
\description{
Combines multiple conditioning taxa into a single context variable
}
\keyword{internal}
