% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualization_dashboards.R
\name{create_interactive_dashboard}
\alias{create_interactive_dashboard}
\title{Create interactive dashboard}
\usage{
create_interactive_dashboard(plots)
}
\description{
Create interactive dashboard
}
\keyword{internal}
