% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualization_networks.R
\name{create_interactive_network}
\alias{create_interactive_network}
\title{Create interactive network visualization}
\usage{
create_interactive_network(network_data, layout, edge_width_scale)
}
\description{
Create interactive network visualization
}
\keyword{internal}
