% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/robust_taxa_indicators.R
\name{create_robust_consensus}
\alias{create_robust_consensus}
\title{Create Robust Consensus Rankings}
\usage{
create_robust_consensus(results, components, top_n)
}
\arguments{
\item{results}{Results from all methods}

\item{components}{Universal information components}

\item{top_n}{Number of top taxa to include}
}
\value{
List of consensus rankings by component
}
\description{
Combines results from multiple methods into consensus rankings
}
\keyword{internal}
