% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualization_dashboards.R
\name{create_sample_profiles_plot}
\alias{create_sample_profiles_plot}
\title{Create sample profiles plot}
\usage{
create_sample_profiles_plot(components, n_samples = 10)
}
\description{
Create sample profiles plot
}
\keyword{internal}
