% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/progress_tracking.R
\name{create_shiny_progress}
\alias{create_shiny_progress}
\title{Create Shiny-based progress tracker}
\usage{
create_shiny_progress(total, title, show_time, show_percent)
}
\description{
Create Shiny-based progress tracker
}
\keyword{internal}
