% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualization_dashboards.R
\name{create_static_dashboard}
\alias{create_static_dashboard}
\title{Create static dashboard layout}
\usage{
create_static_dashboard(plots, n_cols)
}
\description{
Create static dashboard layout
}
\keyword{internal}
