% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualization_networks.R
\name{create_static_network}
\alias{create_static_network}
\title{Create static network visualization}
\usage{
create_static_network(network_data, layout, edge_width_scale)
}
\description{
Create static network visualization
}
\keyword{internal}
