% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/universal_information.R
\name{decompose_to_information_components}
\alias{decompose_to_information_components}
\title{Decompose diversity metrics into universal information components}
\usage{
decompose_to_information_components(
  phyloseq_object,
  diversity_metrics,
  include_phylogenetic,
  include_spatial
)
}
\description{
Decompose diversity metrics into universal information components
}
\keyword{internal}
