% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/information_theory.R
\name{discretize_taxa_abundances}
\alias{discretize_taxa_abundances}
\title{Discretize Taxa Abundances}
\usage{
discretize_taxa_abundances(otu_mat, method, n_bins)
}
\arguments{
\item{otu_mat}{OTU abundance matrix (taxa x samples)}

\item{method}{Discretization method}

\item{n_bins}{Number of discrete bins}
}
\value{
Matrix of discretized abundances
}
\description{
Converts continuous abundance data to discrete categories for mutual information
}
\keyword{internal}
