% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cache_system.R
\docType{data}
\name{.diversityGPT_cache}
\alias{.diversityGPT_cache}
\title{Smart Caching System for diversityGPT}
\format{
An object of class \code{environment} of length 0.
}
\usage{
.diversityGPT_cache
}
\description{
Intelligent caching system for datasets, universal transformations,
and computed results to improve performance and user experience
}
\keyword{datasets}
