% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/universal_transformations.R
\name{estimate_information_components}
\alias{estimate_information_components}
\title{Estimate information components from available metrics}
\usage{
estimate_information_components(source_df, transformation_matrix)
}
\description{
Estimate information components from available metrics
}
\keyword{internal}
