% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shapley_values.R
\name{find_symmetric_taxa}
\alias{find_symmetric_taxa}
\title{Find Symmetric Taxa}
\usage{
find_symmetric_taxa(shapley_values, tolerance = 0.001)
}
\arguments{
\item{shapley_values}{Vector of Shapley values}

\item{tolerance}{Tolerance for considering values equal}
}
\value{
List of symmetric groups
}
\description{
Identifies groups of taxa with similar Shapley values (potential symmetry)
}
\keyword{internal}
