% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shapley_values.R
\name{generate_all_coalitions}
\alias{generate_all_coalitions}
\title{Generate All Possible Coalitions}
\usage{
generate_all_coalitions(n_taxa)
}
\arguments{
\item{n_taxa}{Number of taxa}
}
\value{
List of coalition vectors
}
\description{
Creates all possible subsets of taxa indices
}
\keyword{internal}
