% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bootstrap_uncertainty.R
\name{generate_block_bootstrap_microbiome}
\alias{generate_block_bootstrap_microbiome}
\title{Generate Block Bootstrap Microbiome}
\usage{
generate_block_bootstrap_microbiome(otu_mat, stratify_by)
}
\arguments{
\item{otu_mat}{Original OTU matrix}

\item{stratify_by}{Stratification variable}
}
\value{
Bootstrap OTU matrix
}
\description{
Creates block bootstrap for samples with temporal/spatial structure
}
\keyword{internal}
