% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/consensus_diversity.R
\name{generate_consensus_interpretation}
\alias{generate_consensus_interpretation}
\title{Generate interpretation of consensus results}
\usage{
generate_consensus_interpretation(weights, conflict_analysis, method)
}
\description{
Generate interpretation of consensus results
}
\keyword{internal}
