% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bootstrap_uncertainty.R
\name{generate_parametric_bootstrap_microbiome}
\alias{generate_parametric_bootstrap_microbiome}
\title{Generate Parametric Bootstrap Microbiome}
\usage{
generate_parametric_bootstrap_microbiome(otu_mat)
}
\arguments{
\item{otu_mat}{Original OTU matrix}
}
\value{
Bootstrap OTU matrix
}
\description{
Creates parametric bootstrap sample using negative binomial model
}
\keyword{internal}
