% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cached_analysis.R
\name{generate_universal_cache_key}
\alias{generate_universal_cache_key}
\title{Generate cache key for universal information}
\usage{
generate_universal_cache_key(physeq, groups, include_phylogenetic)
}
\description{
Generate cache key for universal information
}
\keyword{internal}
