% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api_framework.R
\name{get_api_data}
\alias{get_api_data}
\title{Extract Data from API Response}
\usage{
get_api_data(x, component = NULL, ...)
}
\arguments{
\item{x}{API response object}

\item{component}{Optional component to extract}

\item{...}{Additional arguments}
}
\value{
Extracted data
}
\description{
Extract Data from API Response
}
