% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api_framework.R
\name{get_api_error}
\alias{get_api_error}
\title{Get API Error Details}
\usage{
get_api_error(x)
}
\arguments{
\item{x}{API response object}
}
\value{
Error details or NULL
}
\description{
Get API Error Details
}
