% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shapley_values.R
\name{get_complexity_estimate}
\alias{get_complexity_estimate}
\title{Get Computational Complexity Estimate}
\usage{
get_complexity_estimate(method, n_taxa)
}
\arguments{
\item{method}{Calculation method}

\item{n_taxa}{Number of taxa}
}
\value{
Complexity description
}
\description{
Get Computational Complexity Estimate
}
\keyword{internal}
