% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/taxa_drivers.R
\name{identify_phylogenetic_drivers}
\alias{identify_phylogenetic_drivers}
\title{Identify Phylogenetic Drivers}
\usage{
identify_phylogenetic_drivers(
  physeq,
  components,
  top_n,
  method,
  groups,
  normalize
)
}
\arguments{
\item{physeq}{A phyloseq object}

\item{components}{Universal information components}

\item{top_n}{Number of top drivers to return}

\item{method}{Driver identification method}

\item{groups}{Optional grouping variable}

\item{normalize}{Normalize contributions}
}
\value{
Data frame of phylogenetic driver taxa
}
\description{
Identifies taxa that contribute most to phylogenetic diversity patterns
}
\keyword{internal}
