% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/taxa_drivers.R
\name{identify_spatial_drivers}
\alias{identify_spatial_drivers}
\title{Identify Spatial Drivers}
\usage{
identify_spatial_drivers(physeq, components, top_n, method, groups, normalize)
}
\arguments{
\item{physeq}{A phyloseq object}

\item{components}{Universal information components}

\item{top_n}{Number of top drivers to return}

\item{method}{Driver identification method}

\item{groups}{Optional grouping variable}

\item{normalize}{Normalize contributions}
}
\value{
Data frame of spatial driver taxa
}
\description{
Identifies taxa that contribute most to spatial patterns
}
\keyword{internal}
