% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/taxa_indicators.R
\name{identify_spatial_indicators}
\alias{identify_spatial_indicators}
\title{Identify Spatial Indicators}
\usage{
identify_spatial_indicators(
  physeq,
  components,
  top_n,
  method,
  groups,
  normalize
)
}
\arguments{
\item{physeq}{A phyloseq object}

\item{components}{Universal information components}

\item{top_n}{Number of top indicators to return}

\item{method}{Indicator identification method}

\item{groups}{Optional grouping variable}

\item{normalize}{Normalize contributions}
}
\value{
Data frame of spatial indicator taxa
}
\description{
Identifies taxa that serve as indicators for spatial patterns
}
\keyword{internal}
