% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/information_theory.R
\name{information_theory}
\alias{information_theory}
\title{Information Theory Approach to Taxa Indicators}
\description{
Advanced information-theoretic methods for quantifying taxa-component relationships
}
\keyword{information}
\keyword{mutual}
\keyword{theory,}
