% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api_framework.R
\name{is_api_success}
\alias{is_api_success}
\title{Check API Response Status}
\usage{
is_api_success(x)
}
\arguments{
\item{x}{API response object}
}
\value{
Logical indicating success
}
\description{
Check API Response Status
}
