% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shiny_app.R
\name{launch_component_explorer}
\alias{launch_component_explorer}
\title{Launch Component Explorer}
\usage{
launch_component_explorer(
  launch.browser = TRUE,
  port = getOption("shiny.port"),
  host = getOption("shiny.host", "127.0.0.1")
)
}
\arguments{
\item{launch.browser}{Logical; should the app launch in a browser?}

\item{port}{The TCP port that the application should listen on.}

\item{host}{The IPv4 address that the application should listen on.}
}
\description{
Interactive widget with sliders to explore R, E, P, S parameter space
and see how components create all diversity metrics in real-time
}
\examples{
\dontrun{
# Launch component explorer
launch_component_explorer()
}
}
