% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataset_loaders.R
\name{load_curated_mgd}
\alias{load_curated_mgd}
\title{Load dataset from curatedMetagenomicData}
\usage{
load_curated_mgd(
  dataset_info,
  study_name = NULL,
  dryrun = FALSE,
  counts = TRUE,
  ...
)
}
\arguments{
\item{dataset_info}{List containing dataset metadata}

\item{study_name}{Specific study to load}

\item{dryrun}{If TRUE, only show available data}

\item{counts}{Return count data (vs relative abundance)}

\item{...}{Additional arguments passed to curatedMetagenomicData()}
}
\value{
phyloseq object
}
\description{
Load dataset from curatedMetagenomicData
}
\keyword{internal}
