% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataset_loaders.R
\name{load_format_converted}
\alias{load_format_converted}
\title{Load dataset using format converters}
\usage{
load_format_converted(dataset_info, ...)
}
\arguments{
\item{dataset_info}{List containing dataset metadata}

\item{...}{Additional arguments passed to format converters}
}
\value{
phyloseq object
}
\description{
Load dataset using format converters
}
\keyword{internal}
