% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataset_loaders.R
\name{load_microbiome_pkg}
\alias{load_microbiome_pkg}
\title{Load dataset from microbiome package}
\usage{
load_microbiome_pkg(dataset_info, dataset_name = NULL, ...)
}
\arguments{
\item{dataset_info}{List containing dataset metadata}

\item{dataset_name}{Specific dataset to load}

\item{...}{Additional arguments (unused)}
}
\value{
phyloseq object
}
\description{
Load dataset from microbiome package
}
\keyword{internal}
