% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataset_loaders.R
\name{load_precomputed}
\alias{load_precomputed}
\title{Load precomputed universal results}
\usage{
load_precomputed(dataset_info, return_physeq = FALSE, ...)
}
\arguments{
\item{dataset_info}{List containing dataset metadata}

\item{return_physeq}{If TRUE, return phyloseq object if available}

\item{...}{Additional arguments (unused)}
}
\value{
universal_information object or phyloseq object
}
\description{
Load precomputed universal results
}
\keyword{internal}
