% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/performance_optimization.R
\name{load_precomputed_universal}
\alias{load_precomputed_universal}
\title{Load precomputed universal information}
\usage{
load_precomputed_universal(file_path, verify = TRUE)
}
\arguments{
\item{file_path}{Path to precomputed RDS file}

\item{verify}{Check integrity of loaded data}
}
\value{
Universal information object
}
\description{
Load precomputed universal information
}
