% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format_converters.R
\name{merge_qiime2_artifacts}
\alias{merge_qiime2_artifacts}
\title{Merge multiple QIIME2 artifacts into phyloseq}
\usage{
merge_qiime2_artifacts(
  feature_table,
  taxonomy = NULL,
  tree = NULL,
  metadata = NULL
)
}
\arguments{
\item{feature_table}{Path to feature table QZA}

\item{taxonomy}{Optional path to taxonomy QZA}

\item{tree}{Optional path to tree QZA}

\item{metadata}{Optional path to sample metadata TSV}
}
\value{
phyloseq object
}
\description{
Merge multiple QIIME2 artifacts into phyloseq
}
