% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/performance_optimization.R
\name{merge_universal_results}
\alias{merge_universal_results}
\title{Merge universal information results from chunks}
\usage{
merge_universal_results(chunk_results)
}
\description{
Merge universal information results from chunks
}
\keyword{internal}
