% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bootstrap_uncertainty.R
\name{plot.bootstrap_indicators}
\alias{plot.bootstrap_indicators}
\title{Plot Bootstrap Indicators}
\usage{
\method{plot}{bootstrap_indicators}(
  x,
  type = c("confidence_intervals", "uncertainty", "consensus", "method_comparison"),
  component = "richness",
  top_n = 15,
  interactive = FALSE,
  ...
)
}
\arguments{
\item{x}{A bootstrap_indicators object}

\item{type}{Plot type: "confidence_intervals", "uncertainty", "consensus", "method_comparison"}

\item{component}{Component to visualize (default: "richness")}

\item{top_n}{Number of top taxa to show}

\item{interactive}{Create interactive plot using plotly}

\item{...}{Additional arguments passed to plotting functions}
}
\value{
A ggplot2 or plotly object
}
\description{
Creates visualizations of bootstrap uncertainty analysis
}
