% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/meta_analysis.R
\name{plot.meta_analysis}
\alias{plot.meta_analysis}
\title{Plot Meta-Analysis Results}
\usage{
\method{plot}{meta_analysis}(x, type = "forest", component = NULL, ...)
}
\arguments{
\item{x}{Meta-analysis results object}

\item{type}{Plot type: "forest", "heterogeneity", "effects", "studies"}

\item{component}{Component to plot (for forest plots)}

\item{...}{Additional arguments}
}
\description{
Create visualizations of meta-analysis results including forest plots
}
