% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/statistical_validation.R
\name{plot.statistical_validation}
\alias{plot.statistical_validation}
\title{Plot Statistical Validation Results}
\usage{
\method{plot}{statistical_validation}(x, type = "overview", ...)
}
\arguments{
\item{x}{Statistical validation object}

\item{type}{Plot type: "overview", "bootstrap", "cv", "components"}

\item{...}{Additional arguments}
}
\description{
Create visualizations of statistical validation results
}
