% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualization_networks.R
\name{plot.universal_information}
\alias{plot.universal_information}
\title{Plot method for universal_information objects}
\usage{
\method{plot}{universal_information}(x, type = c("network", "components", "quality"), ...)
}
\arguments{
\item{x}{A universal_information object}

\item{type}{Character: plot type ("network", "components", "quality")}

\item{...}{Additional arguments passed to specific plot functions}
}
\description{
Plot method for universal_information objects
}
