% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualization_networks.R
\name{plot.universal_transformation}
\alias{plot.universal_transformation}
\title{Plot method for universal_transformation objects}
\usage{
\method{plot}{universal_transformation}(x, ...)
}
\arguments{
\item{x}{A universal_transformation object}

\item{...}{Additional arguments (unused)}
}
\description{
Plot method for universal_transformation objects
}
