% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/null_model_validation.R
\name{plot.validated_indicators}
\alias{plot.validated_indicators}
\title{Plot Null Model Validation Results}
\usage{
\method{plot}{validated_indicators}(
  x,
  type = c("pvalues", "effects", "null_dist"),
  component = "richness",
  ...
)
}
\arguments{
\item{x}{A validated_indicators object}

\item{type}{Plot type: "pvalues", "effects", "null_dist"}

\item{component}{Which component to plot}

\item{...}{Additional arguments}
}
\description{
Plot Null Model Validation Results
}
