% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bootstrap_uncertainty.R
\name{plot_bootstrap_confidence_intervals}
\alias{plot_bootstrap_confidence_intervals}
\title{Plot Bootstrap Confidence Intervals}
\usage{
plot_bootstrap_confidence_intervals(x, component, top_n, interactive)
}
\arguments{
\item{x}{Bootstrap indicators object}

\item{component}{Component to plot}

\item{top_n}{Number of taxa to show}

\item{interactive}{Use plotly}
}
\value{
Plot object
}
\description{
Plot Bootstrap Confidence Intervals
}
\keyword{internal}
