% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bootstrap_uncertainty.R
\name{plot_consensus_rankings}
\alias{plot_consensus_rankings}
\title{Plot Consensus Rankings}
\usage{
plot_consensus_rankings(x, top_n, interactive)
}
\arguments{
\item{x}{Bootstrap indicators object}

\item{top_n}{Number of taxa to show}

\item{interactive}{Use plotly}
}
\value{
Plot object
}
\description{
Plot Consensus Rankings
}
\keyword{internal}
