% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/taxa_drivers.R
\name{plot_driver_bars}
\alias{plot_driver_bars}
\title{Plot Driver Bars}
\usage{
plot_driver_bars(drivers, top_n, interactive)
}
\arguments{
\item{drivers}{Taxa drivers object}

\item{top_n}{Number of top taxa}

\item{interactive}{Use plotly}
}
\value{
Plot object
}
\description{
Plot Driver Bars
}
\keyword{internal}
