% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/taxa_driver_plots.R
\name{plot_driver_contributions}
\alias{plot_driver_contributions}
\title{Plot Driver Contributions}
\usage{
plot_driver_contributions(drivers, top_n, interactive)
}
\arguments{
\item{drivers}{Taxa drivers object}

\item{top_n}{Number of top taxa}

\item{interactive}{Use plotly}
}
\value{
Contribution plot object
}
\description{
Creates a comprehensive visualization of taxa contributions
}
\keyword{internal}
