% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/taxa_driver_plots.R
\name{plot_driver_network}
\alias{plot_driver_network}
\title{Plot Driver Network}
\usage{
plot_driver_network(drivers, top_n, interactive)
}
\arguments{
\item{drivers}{Taxa drivers object}

\item{top_n}{Number of top taxa per component}

\item{interactive}{Use interactive networkD3}
}
\value{
Network plot object
}
\description{
Creates a network visualization showing relationships between taxa and components
}
\keyword{internal}
