% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/taxa_indicators.R
\name{plot_indicator_bars}
\alias{plot_indicator_bars}
\title{Plot Indicator Bars}
\usage{
plot_indicator_bars(indicators, top_n, interactive)
}
\arguments{
\item{indicators}{Taxa indicators object}

\item{top_n}{Number of top taxa}

\item{interactive}{Use plotly}
}
\value{
Plot object
}
\description{
Plot Indicator Bars
}
\keyword{internal}
