% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/taxa_indicator_plots.R
\name{plot_indicator_contributions}
\alias{plot_indicator_contributions}
\title{Plot Indicator Contributions}
\usage{
plot_indicator_contributions(indicators, top_n, interactive)
}
\arguments{
\item{indicators}{Taxa indicators object}

\item{top_n}{Number of top taxa}

\item{interactive}{Use plotly}
}
\value{
Contribution plot object
}
\description{
Creates a comprehensive visualization of taxa contributions
}
\keyword{internal}
