% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bootstrap_uncertainty.R
\name{plot_method_comparison}
\alias{plot_method_comparison}
\title{Plot Method Comparison}
\usage{
plot_method_comparison(x, component, interactive)
}
\arguments{
\item{x}{Bootstrap indicators object}

\item{component}{Component to plot}

\item{interactive}{Use plotly}
}
\value{
Plot object
}
\description{
Plot Method Comparison
}
\keyword{internal}
