% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/information_theory.R
\name{plot_mi_heatmap}
\alias{plot_mi_heatmap}
\title{Plot Mutual Information Heatmap}
\usage{
plot_mi_heatmap(x, top_n, interactive)
}
\arguments{
\item{x}{Mutual information object}

\item{top_n}{Number of top taxa}

\item{interactive}{Use plotly}
}
\value{
Plot object
}
\description{
Plot Mutual Information Heatmap
}
\keyword{internal}
