% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/information_theory.R
\name{plot_mi_network}
\alias{plot_mi_network}
\title{Plot Mutual Information Network}
\usage{
plot_mi_network(x, top_n, interactive)
}
\arguments{
\item{x}{Mutual information object}

\item{top_n}{Number of top taxa}

\item{interactive}{Use plotly}
}
\value{
Plot object
}
\description{
Plot Mutual Information Network
}
\keyword{internal}
