% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/null_model_plots.R
\name{plot_null_distributions}
\alias{plot_null_distributions}
\title{Plot Null Distributions}
\usage{
plot_null_distributions(validated, component, top_n = 6)
}
\arguments{
\item{validated}{A validated_indicators object}

\item{component}{Which component to plot}

\item{top_n}{Number of top indicators to show (default: 6)}
}
\value{
A ggplot2 object
}
\description{
Shows observed values against null distributions for top indicators
}
\keyword{internal}
