% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/robust_taxa_indicators.R
\name{plot_robust_consensus}
\alias{plot_robust_consensus}
\title{Plot Consensus Rankings}
\usage{
plot_robust_consensus(x, component, top_n, interactive)
}
\arguments{
\item{x}{Robust indicators object}

\item{component}{Component to plot}

\item{top_n}{Number of taxa}

\item{interactive}{Use plotly}
}
\value{
Plot object
}
\description{
Plot Consensus Rankings
}
\keyword{internal}
