% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shapley_values.R
\name{plot_shapley_network}
\alias{plot_shapley_network}
\title{Plot Shapley Network}
\usage{
plot_shapley_network(x, top_n, interactive)
}
\arguments{
\item{x}{Shapley values object}

\item{top_n}{Number of taxa to show}

\item{interactive}{Use plotly}
}
\value{
Plot object
}
\description{
Plot Shapley Network
}
\keyword{internal}
