% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shapley_values.R
\name{plot_shapley_waterfall}
\alias{plot_shapley_waterfall}
\title{Plot Shapley Waterfall Chart}
\usage{
plot_shapley_waterfall(x, component, top_n, interactive)
}
\arguments{
\item{x}{Shapley values object}

\item{component}{Component to plot}

\item{top_n}{Number of taxa to show}

\item{interactive}{Use plotly}
}
\value{
Plot object
}
\description{
Plot Shapley Waterfall Chart
}
\keyword{internal}
