% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/universal_transformations.R
\name{predict_metrics_from_components}
\alias{predict_metrics_from_components}
\title{Predict metrics from information components}
\usage{
predict_metrics_from_components(
  estimated_components,
  target_metrics,
  transformation_matrix,
  quality_threshold
)
}
\description{
Predict metrics from information components
}
\keyword{internal}
