% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/information_theory.R
\name{prepare_component_data}
\alias{prepare_component_data}
\title{Prepare Component Data for Information Theory}
\usage{
prepare_component_data(components, n_samples)
}
\arguments{
\item{components}{Universal information components}

\item{n_samples}{Number of samples expected}
}
\value{
Matrix of component values (samples x components)
}
\description{
Extracts and formats component values for mutual information calculations
}
\keyword{internal}
