% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/consensus_diversity.R
\name{print.consensus_results}
\alias{print.consensus_results}
\title{Print method for consensus results}
\usage{
\method{print}{consensus_results}(x, ...)
}
\arguments{
\item{x}{A consensus_results object}

\item{...}{Additional arguments (unused)}
}
\description{
Print method for consensus results
}
