% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_diversity.R
\name{print.diversity_results}
\alias{print.diversity_results}
\title{Print method for diversity results}
\usage{
\method{print}{diversity_results}(x, ...)
}
\arguments{
\item{x}{A diversity_results object}

\item{...}{Additional arguments (unused)}
}
\description{
Print method for diversity results
}
