% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/robust_taxa_indicators.R
\name{print.robust_taxa_indicators}
\alias{print.robust_taxa_indicators}
\title{Print Method for Robust Taxa Indicators}
\usage{
\method{print}{robust_taxa_indicators}(x, ...)
}
\arguments{
\item{x}{A robust_taxa_indicators object}

\item{...}{Additional arguments (ignored)}
}
\description{
Print Method for Robust Taxa Indicators
}
