% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/taxa_drivers.R
\name{print.taxa_drivers}
\alias{print.taxa_drivers}
\title{Print Method for Taxa Drivers}
\usage{
\method{print}{taxa_drivers}(x, ...)
}
\arguments{
\item{x}{A taxa_drivers object}

\item{...}{Additional arguments (ignored)}
}
\description{
Print Method for Taxa Drivers
}
