% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/taxa_indicators.R
\name{print.taxa_indicators}
\alias{print.taxa_indicators}
\title{Print Method for Taxa Indicators}
\usage{
\method{print}{taxa_indicators}(x, ...)
}
\arguments{
\item{x}{A taxa_indicators object}

\item{...}{Additional arguments (ignored)}
}
\description{
Print Method for Taxa Indicators
}
