% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/universal_information.R
\name{print.universal_information}
\alias{print.universal_information}
\title{Print method for universal_information objects}
\usage{
\method{print}{universal_information}(x, ...)
}
\arguments{
\item{x}{A universal_information object}

\item{...}{Additional arguments (unused)}
}
\description{
Print method for universal_information objects
}
