% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/universal_transformations.R
\name{print.universal_transformation}
\alias{print.universal_transformation}
\title{Print method for universal_transformation objects}
\usage{
\method{print}{universal_transformation}(x, ...)
}
\arguments{
\item{x}{A universal_transformation object}

\item{...}{Additional arguments (unused)}
}
\description{
Print method for universal_transformation objects
}
