% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/information_theory.R
\name{rank_taxa_by_conditional_information}
\alias{rank_taxa_by_conditional_information}
\title{Rank Taxa by Conditional Information}
\usage{
rank_taxa_by_conditional_information(
  cmi_matrix,
  umi_matrix,
  interaction_effects,
  target_taxa
)
}
\arguments{
\item{cmi_matrix}{Conditional MI matrix}

\item{umi_matrix}{Unconditional MI matrix}

\item{interaction_effects}{Interaction effects matrix}

\item{target_taxa}{Target taxa names}
}
\value{
Rankings list
}
\description{
Rank Taxa by Conditional Information
}
\keyword{internal}
