% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/information_theory.R
\name{rank_taxa_by_information}
\alias{rank_taxa_by_information}
\title{Rank Taxa by Information Content}
\usage{
rank_taxa_by_information(mi_matrix, normalized_mi, taxa_names)
}
\arguments{
\item{mi_matrix}{Mutual information matrix}

\item{normalized_mi}{Normalized mutual information matrix}

\item{taxa_names}{Taxa names}
}
\value{
List of ranked taxa for each component
}
\description{
Creates ranked lists of taxa for each component based on information metrics
}
\keyword{internal}
