% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shapley_values.R
\name{rank_taxa_by_shapley_values}
\alias{rank_taxa_by_shapley_values}
\title{Rank Taxa by Shapley Values}
\usage{
rank_taxa_by_shapley_values(shapley_matrix, taxa_names)
}
\arguments{
\item{shapley_matrix}{Matrix of Shapley values}

\item{taxa_names}{Names of taxa}
}
\value{
List of ranked taxa for each component
}
\description{
Creates ranked lists of taxa for each component based on Shapley values
}
\keyword{internal}
