% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/batch_processing.R
\name{retry_failed_datasets}
\alias{retry_failed_datasets}
\title{Retry Failed Datasets}
\usage{
retry_failed_datasets(batch_results, phyloseq_list, ...)
}
\arguments{
\item{batch_results}{Previous batch processing results}

\item{phyloseq_list}{Original list of phyloseq objects}

\item{...}{Additional arguments passed to batch_process_phyloseq}
}
\value{
Updated batch processing results
}
\description{
Retry processing of datasets that failed in previous batch run
}
