% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataset_registry.R
\name{search_datasets}
\alias{search_datasets}
\title{Search datasets by keyword}
\usage{
search_datasets(keyword, fields = c("name", "description", "tags"))
}
\arguments{
\item{keyword}{Search term to find in name, description, or tags}

\item{fields}{Which fields to search in}
}
\value{
Data frame with matching datasets
}
\description{
Search datasets by keyword
}
\examples{
# Search for gut-related datasets
search_datasets("gut")

# Search for time series data
search_datasets("longitudinal")

}
