% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shapley_values.R
\name{shapley_values}
\alias{shapley_values}
\title{Shapley Value Attribution for Taxa Contributions}
\description{
Game-theoretic approach for fair attribution of component values among taxa
}
\keyword{attribution}
\keyword{fair}
\keyword{game}
\keyword{theory,}
