% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/information_theory.R
\name{summarize_conditional_mi}
\alias{summarize_conditional_mi}
\title{Summarize Conditional MI Results}
\usage{
summarize_conditional_mi(
  cmi_matrix,
  umi_matrix,
  interaction_effects,
  conditioning_taxa,
  method
)
}
\arguments{
\item{cmi_matrix}{Conditional MI matrix}

\item{umi_matrix}{Unconditional MI matrix}

\item{interaction_effects}{Interaction effects matrix}

\item{conditioning_taxa}{Conditioning taxa}

\item{method}{Estimation method}
}
\value{
Summary statistics
}
\description{
Summarize Conditional MI Results
}
\keyword{internal}
