% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/information_theory.R
\name{summarize_mutual_information}
\alias{summarize_mutual_information}
\title{Summarize Mutual Information Results}
\usage{
summarize_mutual_information(mi_matrix, normalized_mi, info_gain, method)
}
\arguments{
\item{mi_matrix}{Mutual information matrix}

\item{normalized_mi}{Normalized MI matrix}

\item{info_gain}{Information gain matrix}

\item{method}{Estimation method used}
}
\value{
List of summary statistics
}
\description{
Creates summary statistics and diagnostics
}
\keyword{internal}
