% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/taxa_drivers.R
\name{summarize_taxa_drivers}
\alias{summarize_taxa_drivers}
\title{Summarize Taxa Drivers}
\usage{
summarize_taxa_drivers(drivers)
}
\arguments{
\item{drivers}{Taxa drivers results}
}
\value{
List of summary statistics
}
\description{
Creates summary statistics for taxa driver analysis
}
\keyword{internal}
