% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/taxa_indicators.R
\name{summarize_taxa_indicators}
\alias{summarize_taxa_indicators}
\title{Summarize Taxa Indicators}
\usage{
summarize_taxa_indicators(indicators)
}
\arguments{
\item{indicators}{Taxa indicators results}
}
\value{
List of summary statistics
}
\description{
Creates summary statistics for taxa indicator analysis
}
\keyword{internal}
