% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/taxa_drivers.R
\name{taxa_drivers}
\alias{taxa_drivers}
\title{Taxa Driver Analysis for Information Components}
\description{
Functions to identify which taxa drive the R, E, P, and S information components
}
\keyword{analysis}
\keyword{taxa}
