% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/taxa_indicators.R
\name{taxa_indicators}
\alias{taxa_indicators}
\title{Taxa Indicator Analysis for Information Components}
\description{
Functions to identify which taxa are indicators for the R, E, P, and S information components
}
\keyword{analysis}
\keyword{taxa}
