% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/progress_tracking.R
\name{update_console_progress}
\alias{update_console_progress}
\title{Update console progress}
\usage{
update_console_progress(tracker, message)
}
\description{
Update console progress
}
\keyword{internal}
