% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/progress_tracking.R
\name{update_progress}
\alias{update_progress}
\title{Update progress tracker}
\usage{
update_progress(tracker, current, message = NULL)
}
\arguments{
\item{tracker}{Progress tracker object}

\item{current}{Current step number}

\item{message}{Optional message to display}
}
\description{
Update progress tracker
}
