% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shapley_values.R
\name{validate_shapley_axioms}
\alias{validate_shapley_axioms}
\title{Validate Shapley Axioms}
\usage{
validate_shapley_axioms(shapley_matrix, char_function, taxa_names)
}
\arguments{
\item{shapley_matrix}{Matrix of Shapley values}

\item{char_function}{Characteristic function}

\item{taxa_names}{Names of taxa}
}
\value{
List of axiom validation results
}
\description{
Checks if calculated Shapley values satisfy the fundamental axioms
}
\keyword{internal}
